backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-pymongo==2.3.0
pymongo==4.6.0
PyJWT==2.8.0
werkzeug==3.0.1

